Move unit test to util
